refactor: move test_orchestrator_config_error.py to tests/unit/ (#541)
Move test file from tests/pyats_core/ to tests/unit/pyats_core/ to reduce
placement debt per review comment #9. Add PyATSTestDirs and pyats_test_dirs
fixture to tests/unit/conftest.py (intentional duplication until #541 fully
consolidates the two conftest files).
