+++ tests/unit/conftest.py
@@ -4,7 +4,8 @@
 """Shared fixtures for unit tests."""
 
 import os
-from typing import Any
+from pathlib import Path
+from typing import Any, NamedTuple
 from unittest.mock import AsyncMock, Mock
 
 import pytest
@@ -17,6 +18,15 @@ from nac_test.pyats_core.constants import (
 )
 
 CONTROLLER_ENV_PREFIXES = ("ACI_", "SDWAN_", "CC_", "MERAKI_", "FMC_", "ISE_")
+
+
+class PyATSTestDirs(NamedTuple):
+    """Directory structure for PyATS orchestrator tests."""
+
+    test_dir: Path
+    output_dir: Path
+    merged_file: Path
+
 
 AUTH_SUCCESS = AuthCheckResult(
     success=True,
@@ -125,3 +135,21 @@ def cc_controller_env(monkeypatch: MonkeyPatch) -> None:
     monkeypatch.setenv("CC_URL", "https://cc.test.com")
     monkeypatch.setenv("CC_USERNAME", "admin")
     monkeypatch.setenv("CC_PASSWORD", "test_pass")
+
+
+@pytest.fixture()
+def pyats_test_dirs(tmp_path: Path) -> PyATSTestDirs:
+    """Create standard directory structure for PyATS orchestrator tests.
+
+    Returns:
+        PyATSTestDirs with test_dir, output_dir, and merged_file paths.
+    """
+    test_dir = tmp_path / "tests"
+    test_dir.mkdir()
+    output_dir = tmp_path / "output"
+    output_dir.mkdir()
+    merged_file = output_dir / "merged.yaml"
+    merged_file.write_text("test: data")
+    return PyATSTestDirs(
+        test_dir=test_dir, output_dir=output_dir, merged_file=merged_file
+    )
